# src/TicketManagement.WebUI/OpenAPIs/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/VenueManagement/venue": {
      "post": {
        "tags": [
          "VenueManagement"
        ],
        "summary": "Add new venue.",
        "operationId": "VenueManagement_InsertVenue",
        "requestBody": {
          "x-name": "venue",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Venue"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "."
          }
        }
      },
      "put": {
        "tags": [
          "VenueManagement"
        ],
        "summary": "Update selected venue.",
        "operationId": "VenueManagement_UpdateVenue",
        "requestBody": {
          "x-name": "venue",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Venue"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "."
          }
        }
      }
    },
    "/api/VenueManagement/venue/{venueId}": {
      "delete": {
        "tags": [
          "VenueManagement"
        ],
        "summary": "Delete selected venue.",
        "operationId": "VenueManagement_DeleteVenue",
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "."
          }
        }
      },
      "get": {
        "tags": [
          "VenueManagement"
        ],
        "summary": "Returns selected venue.",
        "operationId": "VenueManagement_GetByIdVenue",
        "parameters": [
          {
            "name": "venueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ".",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Venue"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/VenueManagement/VenueIdByName/{name}": {
      "get": {
        "tags": [
          "VenueManagement"
        ],
        "summary": "Return VenueId by venues name.",
        "operationId": "VenueManagement_GetVenueIdByName",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ".",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BaseEntity": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Venue": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseEntity"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "name",
              "description",
              "address"
            ],
            "properties": {
              "name": {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              "description": {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              "address": {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              "phone": {
                "type": "string",
                "maxLength": 30,
                "nullable": true
              }
            }
          }
        ]
      }
    }
  }
}